requirements.txt
backports-abc==0.4
certifi==2016.9.26
fuzzywuzzy==0.12.0
numpy==1.11.2
Pillow==3.4.1
python-Levenshtein==0.12.0
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.2
Unidecode==0.4.19
websocket-client==0.37.0
tinys3==0.1.12
requests==2.10.0
boto3==1.4.2
